+++ src/vm/comutilnative.h
@@ -123,7 +123,7 @@ public:
     static FCDECL1(int,     GetGeneration, Object* objUNSAFE);
     static FCDECL0(UINT64,  GetSegmentSize);
     static FCDECL0(int,     GetLastGCPercentTimeInGC);
-    static FCDECL1(UINT64,  GetGenerationSize, int gen);
+    static FCDECL1(size_t,  GetGenerationSize, int gen);
     static 
     INT64 QCALLTYPE GetTotalMemory();
